pw.x

Index

QuantumEspressoIO.read_pdos_atm

QuantumEspressoIO.read_pdos_tot

QuantumEspressoIO.read_projwfc_dos

QuantumEspressoIO.read_projwfc_up

Documentation

QuantumEspressoIO.read_pdos_atm — Method

read_pdos_atm(prefix::AbstractString)

Read all the atom-projected PDOS files.

Arguments

prefix::AbstractString: The prefix of the PDOS files (e.g., qe for qe.pdos_atm#1(Si)_wfc#1(s)).

Returns

PDOS as a vector, which is ordered according to the atom index. Each element of the vector is a tuple with the following fields:

atom_index::Int: Index of the atom (1-based).

atom_label::String: Label of the atom.

pdos: A vector of named tuples, each containing:

wfc_index::Int: Index of the wavefunction.

wfc_label::String: Label of the wavefunction.

columns::Matrix{Float64}: The PDOS data matrix.

header::Vector{String}: The header of the PDOS file.

QuantumEspressoIO.read_pdos_tot — Method

read_pdos_tot(io)

Read projwfc.x total PDOS file.

Arguments

prefix::AbstractString: The prefix of the PDOS file (e.g., qe for qe.pdos_tot). If the filename already ends with .pdos_tot, it will be used directly; otherwise, .pdos_tot will be appended to the prefix.

Returns

columns::Matrix{Float64}: A matrix containing the PDOS data.

header::Vector{String}: A vector containing the first line of the file.

QuantumEspressoIO.read_projwfc_dos — Method

read_projwfc_dos(io)

Read projwfc.x total PDOS (e.g. qe.pdos_tot) or atom-projected file (e.g. qe.pdos_atm#1(Si)_wfc#1(s)).

Arguments

io or filename: An IO stream or a filename to read the DOS data from.

Returns

columns::Matrix{Float64}: A matrix containing the DOS data.

header::Vector{String}: A vector containing the first line of the file.

QuantumEspressoIO.read_projwfc_up — Method

read_projwfc_up(io)

Read the projwfc.x output projwfc_up file.

Returns

params: Dictionary of parameters

orbitals: the projected orbitals, each element is a NamedTuple with fields:

atom_index: index of the atom

atom_label: label of the atom

label: label of the orbital, e.g., "3S"

n: principal quantum number

l: azimuthal quantum number

m: magnetic quantum number

projections: the projection data, size: n_kpoint * n_bands * n_orbitals

n_kpoints: number of kpoints

n_bands: number of bands

n_orbitals: number of projected orbitals